MSPSDS-570: commonize views, methods and logic between reporter and question flows
