+++ web/app/views/application/_case_confirmation.html.slim
@@ -1,0 +1,37 @@
+main#main-content.govuk-main-wrapper role="main"
+  .govuk-grid-row
+    .govuk-grid-column-two-thirds
+      .govuk-panel.govuk-panel--confirmation
+        h2.govuk-panel__title
+          = (investigation.is_case ? "Case " : "Question ") + "created"
+        .govuk-panel__body
+          | Your reference number
+          br
+          strong
+            = get_pretty_id(investigation.id)
+      p
+        | You have been sent a confirmation email.
+      hr.govuk-section-break.govuk-section-break--m.govuk-section-break--visible
+  .govuk-grid-row.mspsds-panel
+    .govuk-grid-column-two-thirds
+      h2.govuk-heading-l
+        = "Do you have " + (investigation.is_case ? "report " : "question ") + "correspondance?"
+      p.govuk-body
+        = "If the " + (investigation.is_case ? "report " : "question ")
+        = "came by email, or you've notes / transcript from a call, add them now."
+      p.govuk-body
+        // TODO MSPSDS-516 implement add reporter correspondance functionality
+        a.govuk-link.mspsds-block-link href="/flows/add-correspondance/01-add-correspondance-context"
+          = "Add " + (investigation.is_case ? "reporter " : "questioner ") + "correspondance"
+      hr.govuk-section-break.govuk-section-break--m.govuk-section-break--visible
+  .govuk-grid-row.mspsds-panelq
+    .govuk-grid-column-two-thirds
+      h2.govuk-heading-l
+        | If you're on the phone right now
+      p.govuk-body
+        | or, if correspondance isn't available right now, don't forget to add it later.
+      p.govuk-body
+        = link_to(\
+          "View the " + (investigation.is_case ? "case " : "question ") + "you've just created",\
+          investigation_path(investigation),\
+          class: "govuk-link mspsds-block-link ")
