+++ web/app/views/investigations/report/type.html.slim
@@ -1,59 +1,4 @@
 - content_for :page_title, "Report unsafe product - Market Surveillance & Product Safety"
 - content_for :after_header do
   = link_to "Back to list", investigations_path, class: "govuk-back-link"
-main#main-content.govuk-main-wrapper role="main"
-  h1.govuk-heading-xl
-    span.govuk-caption-xl Report an unsafe product
-    | Who’s reporting?
-  .govuk-grid-row
-    .govuk-grid-column-two-thirds
-      = form_with(model: @reporter, local: true, url: wizard_path, method: :put) do |form|
-        - if @reporter.errors.any?
-          #error_explanation
-            h2.govuk-heading-l.govuk-error-message
-              = pluralize(@reporter.errors.count, "error ")
-              | prohibited this case from being saved:
-            ul.govuk-list.govuk-error-message
-              - @reporter.errors.full_messages.each do |message|
-                li
-                  = message
-        fieldset.govuk-fieldset
-          legend.govuk-fieldset__legend.govuk-fieldset__legend--l
-            | Reporter type?
-          .govuk-form-group class=("govuk-form-group--error" if @reporter.errors[:reporter_type].any?)
-            .govuk-radios.govuk-radios--conditional data-module="radios"
-              .govuk-radios__item
-                = form.radio_button :reporter_type, "Consumer", class: "govuk-radios__input"
-                = form.label :reporter_type, "Consumer", class: "govuk-label govuk-radios__label"
-                  | Consumer
-              .govuk-radios__item
-                = form.radio_button :reporter_type, "Business", class: "govuk-radios__input"
-                = form.label :reporter_type, "Business", class: "govuk-label govuk-radios__label"
-                  | Business
-              .govuk-radios__item
-                = form.radio_button :reporter_type, "Local authority (Trading standards)",\
-                        class: "govuk-radios__input"
-                = form.label :reporter_type, "Local authority (Trading standards)",\
-                        class: "govuk-label govuk-radios__label"
-                  | Local authority (Trading standards)
-              .govuk-radios__item
-                = form.radio_button :reporter_type, "Trade body", class: "govuk-radios__input"
-                = form.label :reporter_type, "Trade body", class: "govuk-label govuk-radios__label"
-                  | Trade body
-              .govuk-radios__item
-                = form.radio_button :reporter_type, "Other MSA", class: "govuk-radios__input"
-                = form.label :reporter_type, "Other MSA", class: "govuk-label govuk-radios__label"
-                  | Other MSA
-              .govuk-radios__item
-                = form.radio_button :reporter_type, "Other", class: "govuk-radios__input",\
-                        "data-aria-controls": "conditional-reporter-type-other"
-                = form.label :reporter_type, "Other", class: "govuk-label govuk-radios__label"
-                  | Other
-              #conditional-reporter-type-other.govuk-radios__conditional.govuk-radios__conditional--hidden
-                .govuk-form-group
-                  span#other-reporter-hint.govuk-hint [class="govuk-!-width-two-thirds"]
-                    | Please give details
-                  = form.text_field :other_reporter, class: "govuk-input govuk-!-width-two-thirds",\
-                    "aria-describedby": "other-reporter-hint"
-        .govuk-form-group
-          = form.submit "Continue", class: "govuk-button"
+= render "reporter_type_selection", reporter: @reporter, is_case: @investigation.is_case
